requirements.txt
# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Databases
pymongo>=4.7.0
redis>=5.0.0
neo4j>=5.20.0
pymilvus>=2.4.0

# Embeddings
sentence-transformers>=3.0.0
torch>=2.0.0

# Validation & config
pydantic>=2.0.0
python-dotenv>=1.0.0

# Utilities
numpy>=1.26.0